requirements.txt
mmcv==1.7.0
timm==0.6.12
accelerate==0.25.0
tensorboard
tensorboardX
transformers==4.36.1
sentencepiece~=0.1.99
ftfy
beautifulsoup4
protobuf==3.20.2
gradio==4.1.1
yapf==0.40.1
opencv-python
bs4
einops
xformers==0.0.19
optimum
peft
came-pytorch